+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Add `Vector3f.isEqualTo(Vector3f)` method that can be invoked even if the `vecmath.isObjectEqualsImplemented`
+  compile-time constant is set to `false`.
+
 ### [v0.03](https://github.com/realityforge/vecmath/tree/v0.03) (2020-10-04)
 [Full Changelog](https://github.com/realityforge/vecmath/compare/v0.02...v0.03)